+++ src/pages/Portfolio/LendingStats/LendingStatsPage.vue
@@ -32,7 +32,6 @@
 						noun="country"
 						:not-lent-to="countriesNotLentTo"
 						:lent-to="countriesLentTo"
-						:total="totalCountries"
 						item-key="isoCode"
 						unlent-url="/lend/countries-not-lent"
 						section-id="lend-stat-countries"
@@ -66,7 +65,6 @@
 						noun="Lending Partner"
 						:not-lent-to="partnersNotLentTo"
 						:lent-to="partnersLentTo"
-						:total="totalPartners"
 						query="partner"
 						section-id="lend-stat-fieldpartner"
 						show-more-id="show-more-fieldpartner"
@@ -88,8 +86,6 @@
 
 <script>
 import _differenceBy from 'lodash/differenceBy';
-import _get from 'lodash/get';
-import _map from 'lodash/map';
 import _sortBy from 'lodash/sortBy';
 import lendingStatsQuery from '@/graphql/query/myLendingStats.graphql';
 import userAchievementsProgress from '@/graphql/query/userAchievementsProgress.graphql';
@@ -120,14 +116,12 @@ export default {
 		return {
 			countriesLentTo: [],
 			countriesNotLentTo: [],
-			totalCountries: 0,
 			sectorsLentTo: [],
 			sectorsNotLentTo: [],
 			activitiesLentTo: [],
 			activitiesNotLentTo: [],
 			partnersLentTo: [],
 			partnersNotLentTo: [],
-			totalPartners: 0,
 			userId: null,
 			allAchievements: [],
 		};
@@ -136,23 +130,27 @@ export default {
 		query: lendingStatsQuery,
 		preFetch: true,
 		result({ data }) {
-			const allCountries = _sortBy(_map(_get(data, 'lend.countryFacets'), 'country'), 'name');
-			this.countriesLentTo = _sortBy(_get(data, 'my.lendingStats.countriesLentTo'), 'name');
-			this.countriesNotLentTo = _differenceBy(allCountries, this.countriesLentTo, 'isoCode');
-			this.totalCountries = allCountries.length;
+			const countriesWhereKivaDoesntWork = data?.general?.excluded_countries?.value ?? '';
+			// reduce array of country objects to country property without using lodash
+			const countriesReduced = (data?.lend?.countryFacets ?? []).map(country => country.country);
+			// eslint-disable-next-line max-len
+			const filteredCountries = countriesReduced.filter(country => !countriesWhereKivaDoesntWork.includes(country.isoCode));
 
-			const allSectors = _sortBy(_get(data, 'general.kivaStats.sectors'), 'name');
-			this.sectorsLentTo = _sortBy(_get(data, 'my.lendingStats.sectorsLentTo'), 'name');
+			const allCountries = _sortBy(filteredCountries, 'name');
+			this.countriesLentTo = _sortBy(data?.my?.lendingStats?.countriesLentTo ?? [], 'name');
+			this.countriesNotLentTo = _differenceBy(allCountries, this.countriesLentTo, 'isoCode');
+
+			const allSectors = _sortBy(data?.general?.kivaStats?.sectors ?? [], 'name');
+			this.sectorsLentTo = _sortBy(data?.my?.lendingStats?.sectorsLentTo ?? [], 'name');
 			this.sectorsNotLentTo = _differenceBy(allSectors, this.sectorsLentTo, 'id');
 
-			const allActivities = _sortBy(_get(data, 'general.kivaStats.activities'), 'name');
-			this.activitiesLentTo = _sortBy(_get(data, 'my.lendingStats.activitiesLentTo'), 'name');
+			const allActivities = _sortBy(data?.general?.kivaStats?.activities ?? [], 'name');
+			this.activitiesLentTo = _sortBy(data?.my?.lendingStats?.activitiesLentTo, 'name');
 			this.activitiesNotLentTo = _differenceBy(allActivities, this.activitiesLentTo, 'id');
 
-			const allPartners = _sortBy(_get(data, 'general.partners.values'), 'name');
-			this.partnersLentTo = _sortBy(_get(data, 'my.lendingStats.partnersLentTo'), 'name');
+			const allPartners = _sortBy(data?.general?.partners?.values ?? [], 'name');
+			this.partnersLentTo = _sortBy(data?.my?.lendingStats?.partnersLentTo, 'name');
 			this.partnersNotLentTo = _differenceBy(allPartners, this.partnersLentTo, 'id');
-			this.totalPartners = _get(data, 'general.partners.totalCount');
 
 			this.userId = data?.my?.userAccount?.id;
 		},